+++ runtime/integration-tests/src/generic/utils/genesis.rs
@@ -1,7 +1,7 @@
 use frame_support::traits::GenesisBuild;
 use sp_runtime::Storage;
 
-use crate::generic::env::Config;
+use crate::generic::runtime::Runtime;
 
 #[derive(Default)]
 pub struct Genesis {
@@ -9,7 +9,7 @@ pub struct Genesis {
 }
 
 impl Genesis {
-	pub fn add<T: Config>(mut self, builder: impl GenesisBuild<T>) -> Genesis {
+	pub fn add<T: Runtime>(mut self, builder: impl GenesisBuild<T>) -> Genesis {
 		builder.assimilate_storage(&mut self.storage).unwrap();
 		self
 	}
